Move RpmTool to tests entirely
It's no longer used in pulp_rpm proper
