+++ pulp_rpm/tests/functional/utils.py
@@ -16,6 +16,8 @@ import createrepo_c as cr
 import pyzstd
 import requests
 
+from pulpcore.plugin.exceptions import InvalidSignatureError
+
 from pulp_rpm.tests.functional.constants import (
     LEGACY_SIGNING_KEY,
     PACKAGES_DIRECTORY,
@@ -285,3 +287,44 @@ class RepositoryBuilder:
                 writer.add_pkg(cr_pkg)
 
         return RemoteRepository(url=f"file://{repo_dir.absolute()}")
+
+
+class RpmTool:
+    """A wrapper around the rpm CLI for verifying RPM signatures in tests."""
+
+    INVALID_SIGNATURE_ERROR_MSG = "Signature is invalid or pubkey is unreachable"
+    UNKNOWN_ERROR_MSG = "Some unknown error occurred"
+    UNSIGNED_ERROR_MSG = "The package is not signed"
+
+    def __init__(self, root: Optional[Path] = None):
+        completed_process = subprocess.run(
+            ["which", "rpmsign"], stdout=subprocess.PIPE, stderr=subprocess.PIPE
+        )
+        if completed_process.returncode != 0:
+            raise RuntimeError("Rpm cli tool is not installed on your system.")
+
+        self.opts = ["--root", str(root.absolute())] if root else []
+
+    def import_pubkey_file(self, pubkey_file: str):
+        """Import an ASCII-armored public key file into the rpm keyring."""
+        cmd = ("rpm", *self.opts, "--import", pubkey_file)
+        completed_process = subprocess.run(cmd, capture_output=True)
+        if completed_process.returncode != 0:
+            raise RuntimeError(
+                f"Could not import public key into rpm-tool:\n{completed_process.stderr.decode()}"
+            )
+
+    def verify_signature(self, rpm_package_file: Path):
+        """Verify that an RPM package is signed by one of the imported public keys."""
+        cmd = ("rpm", *self.opts, "--checksig", str(rpm_package_file.resolve()))
+        completed_process = subprocess.run(cmd, capture_output=True)
+        stdout = completed_process.stdout.decode()
+        stderr = completed_process.stderr.decode()
+        output = f"\nstdout: {stdout}\nstderr: {stderr}"
+        if completed_process.returncode != 0:
+            if "SIGNATURES NOT OK" in stdout:
+                raise InvalidSignatureError(f"{RpmTool.INVALID_SIGNATURE_ERROR_MSG}: {output}")
+            raise TypeError(f"{RpmTool.UNKNOWN_ERROR_MSG}: {output}")
+        elif "signatures" not in output:
+            raise InvalidSignatureError(f"{RpmTool.UNSIGNED_ERROR_MSG}: {output}")
+        return True
